Use a base test decorated with a collection to force tests to run synchronously
This ensures that we don't have multiple servers running at a time (which causes flaky tests)
