+++ tests/DfE.FindInformationAcademiesTrusts.IntegrationTests/AuthenticationHandlers/UnauthorisedUserAuthenticationHandler.cs
@@ -1,0 +1,30 @@
+using System.Security.Claims;
+using System.Text.Encodings.Web;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.Extensions.Options;
+
+namespace DfE.FindInformationAcademiesTrusts.IntegrationTests.AuthenticationHandlers;
+
+public class UnauthorisedUserAuthenticationHandler(
+    IOptionsMonitor<AuthenticationSchemeOptions> options,
+    ILoggerFactory logger,
+    UrlEncoder encoder)
+    : AuthenticationHandler<AuthenticationSchemeOptions>(options, logger, encoder)
+{
+    public const string AuthenticationScheme = "UnauthorisedUser";
+
+    /// <summary>
+    /// The user authenticates successfully but doesn't have the "User.Role.Authorised" role
+    /// </summary>
+    /// <returns></returns>
+    protected override Task<AuthenticateResult> HandleAuthenticateAsync()
+    {
+        var claims = new[] { new Claim(ClaimTypes.Name, "Test user") };
+        var identity = new ClaimsIdentity(claims, "Test");
+        var principal = new ClaimsPrincipal(identity);
+        var ticket = new AuthenticationTicket(principal, AuthenticationScheme);
+
+        var result = AuthenticateResult.Success(ticket);
+        return Task.FromResult(result);
+    }
+}
