+++ tests/DfE.FindInformationAcademiesTrusts.IntegrationTests/Base/BaseIntegrationTest.cs
@@ -1,0 +1,7 @@
+namespace DfE.FindInformationAcademiesTrusts.IntegrationTests.Base;
+
+[Trait("IntegrationTest", "true")]
+[Collection(nameof(UseWebApplicationFactory))]
+public abstract class BaseIntegrationTest
+{
+}
