+++ tests/DfE.FindInformationAcademiesTrusts.IntegrationTests/Base/UseWebApplicationFactory.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Mvc.Testing;
+
+namespace DfE.FindInformationAcademiesTrusts.IntegrationTests.Base;
+
+[CollectionDefinition(nameof(UseWebApplicationFactory))]
+public class UseWebApplicationFactory : ICollectionFixture<WebApplicationFactory<Program>>
+{
+    // This class has no code, and is never created. Its purpose is simply
+    // to be the place to apply [CollectionDefinition] and all the
+    // ICollectionFixture<> interfaces.
+}
